+++ src/main/resources/db/migration/V3__Monk_Ecom_Script.sql
@@ -1,5 +1,5 @@
 CREATE TABLE Coupon_For_Cart (
-    coupon_id INT PRIMARY KEY,
+    coupon_id BIGINT PRIMARY KEY,
     threshold NUMERIC(9,2),
     discount NUMERIC(9,2),
     description VARCHAR(255)
